+++ app/assets/javascripts/reducers/stories.js
@@ -20,6 +20,10 @@ const storiesReducer = (state = initialState, action) => {
         updateIfSameId(action.story.id, (story) => {
           return updateStory(story, action.story);
         }));
+    case actionTypes.DELETE_STORY_SUCCESS:
+      return state.filter(
+        story => story.id !== action.id
+      );
     default:
       return state;
   };